refactor: 트랜잭션 로직 분리
