+++ src/main/java/chess/infra/db/DBConnectionPool.java
@@ -1,4 +1,4 @@
-package chess.infra;
+package chess.infra.db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
