+++ src/main/java/chess/infra/db/TransactionFunction.java
@@ -1,0 +1,9 @@
+package chess.infra.db;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface TransactionFunction<T> {
+    T execute(Connection conn) throws SQLException;
+}
