+++ src/main/java/chess/service/GameService.java
@@ -10,7 +10,8 @@ import chess.entity.PieceEntity;
 import chess.entity.game.GameEntity;
 import chess.entity.game.repository.GameRepository;
 import chess.entity.piece.repository.PieceRepository;
-import chess.infra.DBConnectionPool;
+import chess.infra.db.DBConnectionPool;
+import chess.infra.db.TransactionManager;
 
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -28,10 +29,7 @@ public class GameService {
     }
 
     public ChessGame createNewGame() throws SQLException {
-        Connection conn = DBConnectionPool.getConnection();
-        try {
-            conn.setAutoCommit(false);
-
+        return TransactionManager.cud(DBConnectionPool.getConnection(), conn -> {
             pieceRepository.deleteAll(conn);
             gameRepository.deleteAll(conn);
 
@@ -43,56 +41,36 @@ public class GameService {
                 Piece piece = entry.getValue();
                 pieceRepository.add(conn, new PieceEntity(gameId, position, piece));
             }
-            conn.commit();
-            return findGameByIdAndTurn(gameId, turn);
-        } catch (Exception exception) {
-            conn.rollback();
-            throw new RuntimeException(exception);
-        } finally {
-            conn.setAutoCommit(true);
-            DBConnectionPool.releaseConnection(conn);
-        }
+            return findGameByIdAndTurn(conn, gameId, turn);
+        });
     }
 
-    private ChessGame findGameByIdAndTurn(Long gameId, Turn turn) {
-        Connection conn = DBConnectionPool.getConnection();
-        try {
+    private ChessGame findGameByIdAndTurn(final Connection connection, final Long gameId, final Turn turn) {
+        return TransactionManager.read(connection, conn -> {
             Map<Position, Piece> board = new HashMap<>();
             List<PieceEntity> pieceEntities = pieceRepository.findByGameId(conn, gameId);
             for (PieceEntity pieceEntity : pieceEntities) {
                 Piece piece = new Piece(pieceEntity.getId(), pieceEntity.getType(), pieceEntity.getColor());
                 Position position = Position.of(pieceEntity.getFile(), pieceEntity.getRank());
-
                 board.put(position, piece);
             }
 
             return new ChessGame(gameId, new ChessBoard(board), turn);
-        } catch (Exception exception) {
-            throw new RuntimeException(exception);
-        } finally {
-            DBConnectionPool.releaseConnection(conn);
-        }
+        });
     }
 
     public ChessGame loadGame() {
-        Connection conn = DBConnectionPool.getConnection();
-        try {
-            GameEntity gameEntity = gameRepository.findLastGame(conn)
-                    .orElseThrow(() -> new IllegalArgumentException("가장 최근 플레이한 게임이 존재하지 않습니다."));
+        return TransactionManager.read(DBConnectionPool.getConnection(), conn -> {
+            GameEntity gameEntity = gameRepository.findLastGame(conn).orElseThrow(
+                    () -> new IllegalArgumentException("가장 최근 플레이한 게임이 존재하지 않습니다."));
             List<PieceEntity> pieceEntities = pieceRepository.findByGameId(conn, gameEntity.getId());
 
             ChessBoard chessBoard = createBoardByPieceEntities(pieceEntities);
-
-            DBConnectionPool.releaseConnection(conn);
             return new ChessGame(gameEntity.getId(), chessBoard, gameEntity.getTurn());
-        } catch (Exception exception) {
-            throw new RuntimeException(exception);
-        } finally {
-            DBConnectionPool.releaseConnection(conn);
-        }
+        });
     }
 
-    private ChessBoard createBoardByPieceEntities(List<PieceEntity> pieceEntities) {
+    private ChessBoard createBoardByPieceEntities(final List<PieceEntity> pieceEntities) {
         Map<Position, Piece> board = new HashMap<>();
         for (PieceEntity pieceEntity : pieceEntities) {
             Piece piece = new Piece(pieceEntity.getId(), pieceEntity.getType(), pieceEntity.getColor());
@@ -105,30 +83,17 @@ public class GameService {
     }
 
     public void updateGame(final Long gameId, final Turn turn, final Long pieceId, final Position target) throws SQLException {
-        Connection conn = DBConnectionPool.getConnection();
-        try {
-            conn.setAutoCommit(false);
+        TransactionManager.cud(DBConnectionPool.getConnection(), conn -> {
             gameRepository.updateTurnById(conn, gameId, turn.now());
             pieceRepository.updatePositionById(conn, pieceId, target.file(), target.rank());
-
-            conn.commit();
-            DBConnectionPool.releaseConnection(conn);
-        } catch (Exception exception) {
-            throw new RuntimeException(exception);
-        } finally {
-            conn.setAutoCommit(true);
-            DBConnectionPool.releaseConnection(conn);
-        }
+            return null;
+        });
     }
 
-    public void deleteLatestGame(final Long gameId) {
-        Connection conn = DBConnectionPool.getConnection();
-        try {
+    public void deleteLatestGame(final Long gameId) throws SQLException {
+        TransactionManager.cud(DBConnectionPool.getConnection(), conn -> {
             gameRepository.deleteById(conn, gameId);
-        } catch (Exception exception) {
-            throw new RuntimeException(exception);
-        } finally {
-            DBConnectionPool.releaseConnection(conn);
-        }
+            return null;
+        });
     }
 }
